+++ src/providers/Fractal/types/ethers.ts
@@ -1,0 +1,6 @@
+import { ethers } from 'ethers';
+
+export type Providers =
+  | ethers.providers.Web3Provider
+  | ethers.providers.JsonRpcProvider
+  | ethers.providers.BaseProvider;
